+++ src/Ksuid.php
@@ -43,21 +43,6 @@ class Ksuid
         return new self;
     }
 
-    public static function fromString($string)
-    {
-        $decoded = (new Base62)->decode($string);
-        return self::fromBytes($decoded);
-    }
-
-    public static function fromBytes($bytes)
-    {
-        $bytes = ltrim($bytes, "\0x00");
-        $timestamp = substr($bytes, 0, self::TIMESTAMP_SIZE);
-        $timestamp = unpack("Nuint", $timestamp);
-        $payload = substr($bytes, self::TIMESTAMP_SIZE, self::PAYLOAD_SIZE);
-        return new self($payload, $timestamp["uint"]);
-    }
-
     public function bytes()
     {
         return pack("N", $this->timestamp) . $this->payload;